requirements.txt
litellm>=1.40.0
gspread>=6.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
weasyprint>=62.0
jinja2>=3.1.0
pydantic>=2.7.0
python-dotenv>=1.0.0
typer[all]>=0.12.0
apscheduler>=3.10.0
tenacity>=8.3.0
pyyaml>=6.0.0
rich>=13.7.0
serpapi>=1.0.0
